+++ types-override/http/accepts.d.ts
@@ -2,7 +2,7 @@ import type {AcceptsOptions, AcceptsResult} from '../ergo.js';
 
 declare function accepts(
   options?: AcceptsOptions
-): ((req: {
+): ((req?: {
   headers?: Record<string, string | string[] | undefined>;
 }) => AcceptsResult | {response: {statusCode: 406; detail: string}}) & {
   readonly setPath: 'accepts';