+++ types-override/http/authorization.d.ts
@@ -2,7 +2,7 @@ import type {AuthorizationOptions, AuthorizationResult} from '../ergo.js';
 
 declare function authorization(
   options?: AuthorizationOptions
-): ((req: {
+): ((req?: {
   headers?: {authorization?: string};
 }) => Promise<
   AuthorizationResult | {response: {statusCode: number; headers?: [string, string][]}}