+++ types-override/http/cookie.d.ts
@@ -2,6 +2,6 @@ import type {CookieOptions, CookieJar} from '../ergo.js';
 
 declare function cookie(
   options?: CookieOptions
-): ((req: {headers?: {cookie?: string}}) => CookieJar) & {readonly setPath: 'cookies'};
+): ((req?: {headers?: {cookie?: string}}) => CookieJar) & {readonly setPath: 'cookies'};
 
 export default cookie;