+++ types-override/http/url.d.ts
@@ -1,5 +1,5 @@
 import type {UrlResult} from '../ergo.js';
 
-declare function url(): ((req: {url?: string}) => UrlResult) & {readonly setPath: 'url'};
+declare function url(): ((req?: {url?: string}) => UrlResult) & {readonly setPath: 'url'};
 
 export default url;