+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.0] - 2026-06-13
+
 ### Added
 
 - **Intrinsic `setPath` on built-in middleware factory return values.** (#153)